fix layout quanlyuser
