+++ .idea/webContexts.xml
@@ -10,6 +10,7 @@
         <entry key="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/admin_page/quanlyCategory.jsp" value="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/admin_page" />
         <entry key="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/admin_page/quanlyProduct.jsp" value="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/admin_page" />
         <entry key="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/admin_page/quanlynhaphang.jsp" value="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/admin_page" />
+        <entry key="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/admin_page/quanlyuser.jsp" value="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/admin_page" />
         <entry key="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/blog-details.jsp" value="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp" />
         <entry key="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/gio-hang.jsp" value="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp" />
         <entry key="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp/index.jsp" value="file://$PROJECT_DIR$/WebBanThuocBaoVeThucVat/src/main/webapp" />
