+++ WebBanThuocBaoVeThucVat/src/main/java/controller/Admin/DeleteProduct.java
@@ -1,7 +1,8 @@
-package controller;
+package controller.Admin;
 
-import Service.CategoryService;
+import Service.ProductsService;
 
+import javax.lang.model.element.Name;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -9,8 +10,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "DeleteCategory", value = "/deleteCate")
-public class DeleteCategory extends HttpServlet {
+@WebServlet(name = "DeleteProduct",value = "/deletePro")
+public class DeleteProduct extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         doPost(req, resp);
@@ -18,12 +19,13 @@ public class DeleteCategory extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String cateID =req.getParameter("cateID");
-        int intCateID=0;
-        if((cateID!= null) && (!cateID.isEmpty()) ){
-            intCateID= Integer.parseInt(cateID);
+        String proID = req.getParameter("proID");
+        String page = req.getParameter("page");
+        int proIDint= 0;
+        if(proID!= null && !proID.isEmpty()){
+            proIDint=Integer.parseInt(proID);
         }
-        CategoryService.getInstance().deleteCategory(intCateID);
-        resp.sendRedirect("./maCategory");
+        ProductsService.getInstance().deleteProduct(proIDint);
+        resp.sendRedirect("./maProduct?proID="+page);
     }
 }
