+++ WebBanThuocBaoVeThucVat/src/main/java/controller/Admin/DeleteUser.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.Admin;
 
 import Service.UserService;
