+++ WebBanThuocBaoVeThucVat/src/main/java/controller/Admin/EditCategory.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.Admin;
 
 import Service.CategoryService;
 
@@ -9,19 +9,25 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "InsertCategory",value = "/insertCate")
-public class InsertCategory extends HttpServlet {
+@WebServlet(name = "EditCategory",value = "/editCategory")
+public class EditCategory extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         req.setCharacterEncoding("UTF-8");
         doPost(req, resp);
+
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         req.setCharacterEncoding("UTF-8");
-        String nameCate= req.getParameter("nameCate");
-        CategoryService.getInstance().insertCategory(nameCate);
+        String categoryName= req.getParameter("categoryName");
+        String categoryId = req.getParameter("categoryId");
+        int intCategoryId=0;
+        if(categoryId!= null && !categoryId.isEmpty()){
+            intCategoryId =Integer.parseInt(categoryId);
+        }
+        CategoryService.getInstance().updateCategory(categoryName,intCategoryId);
         resp.sendRedirect("./maCategory");
     }
 }
