+++ WebBanThuocBaoVeThucVat/src/main/java/controller/Admin/InsertCategory.java
@@ -1,8 +1,7 @@
-package controller;
+package controller.Admin;
 
-import Service.ProductsService;
+import Service.CategoryService;
 
-import javax.lang.model.element.Name;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -10,22 +9,19 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebServlet(name = "DeleteProduct",value = "/deletePro")
-public class DeleteProduct extends HttpServlet {
+@WebServlet(name = "InsertCategory",value = "/insertCate")
+public class InsertCategory extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        req.setCharacterEncoding("UTF-8");
         doPost(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String proID = req.getParameter("proID");
-        String page = req.getParameter("page");
-        int proIDint= 0;
-        if(proID!= null && !proID.isEmpty()){
-            proIDint=Integer.parseInt(proID);
-        }
-        ProductsService.getInstance().deleteProduct(proIDint);
-        resp.sendRedirect("./maProduct?proID="+page);
+        req.setCharacterEncoding("UTF-8");
+        String nameCate= req.getParameter("nameCate");
+        CategoryService.getInstance().insertCategory(nameCate);
+        resp.sendRedirect("./maCategory");
     }
 }
