+++ WebBanThuocBaoVeThucVat/src/main/java/controller/Admin/InsertProduct.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.Admin;
 
 import CKeditor.HtmlUtils;
 import Service.CategoryService;
