+++ WebBanThuocBaoVeThucVat/src/main/java/controller/Admin/InsertUser.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.Admin;
 
 import Service.UserService;
 import bean.User;
