+++ WebBanThuocBaoVeThucVat/src/main/java/dao/UserDAO.java
@@ -176,23 +176,18 @@ public class UserDAO extends AbstractDao<User> {
                         .collect(Collectors.toList()));
         return users;
     }
-    public static List<User> listOfRoleWithSearch(int role, int index, String search) {
+    public static List<User> listOfRoleWithSearch(int role) {
         List<User> users = JDBIConnector.getJdbi().withHandle(handle ->
-                handle.createQuery("SELECT id, role, user_name, password, phone, email"+
-                                ",sur_name,last_name,hash\n" +
-                                "FROM users\n" +
-                                "WHERE role=? AND (last_name LIKE ? OR user_name LIKE ?)\n" +
-                                "ORDER BY id\n" +
-                                "LIMIT 5\n" +
-                                "OFFSET ? ")
+                handle.createQuery("SELECT id, role, user_name, password, phone, email, sur_name, last_name, hash " +
+                                "FROM users " +
+                                "WHERE role = ? " +
+                                "ORDER BY id")
                         .bind(0, role)
-                        .bind(1, "%" + search + "%")
-                        .bind(2, "%" + search + "%")
-                        .bind(3, (index - 1) * 5)
                         .mapToBean(User.class)
                         .collect(Collectors.toList()));
         return users;
     }
+
 
     @Override
     public boolean selectModel(int id) {
